# backend/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Consultation API",
      "version": "1.0.0",
      "description": "API for managing medical consultations"
    },
    "servers": [
      {
        "url": "http://localhost:5000",
        "description": "Local development server"
      }
    ],
    "paths": {
...
    "/api/notification": {
    "get": {
      "summary": "Get all notifications",
      "description": "Retrieves all notifications ordered by creation date",
      "tags": ["Notifications"],
      "responses": {
        "200": {
          "description": "List of notifications",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer", "example": 1 },
                    "provider_id": { "type": "integer", "example": 101 },
                    "type": { "type": "string", "example": "system" },
                    "message": { "type": "string", "example": "System update at 5 PM" },
                    "status": { "type": "string", "example": "unread" },
                    "createdAt": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          }
        },
        "500": {
          "description": "Internal server error"
        }
      }
    }
  },
  "/api/notification/{id}": {
    "get": {
      "summary": "Get a single notification",
      "description": "Retrieves a notification by ID",
      "tags": ["Notifications"],
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "required": true,
          "schema": { "type": "integer" },
          "description": "ID of the notification"
        }
      ],
      "responses": {
        "200": {
          "description": "Notification details",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer", "example": 1 },
                  "provider_id": { "type": "integer", "example": 101 },
                  "type": { "type": "string", "example": "system" },
                  "message": { "type": "string", "example": "System update at 5 PM" },
                  "status": { "type": "string", "example": "unread" },
                  "createdAt": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        },
        "404": {
          "description": "Notification not found"
        },
        "500": {
          "description": "Internal server error"
        }
      }
    },
    "delete": {
      "summary": "Delete a notification",
      "description": "Permanently removes a notification from the system",
      "tags": ["Notifications"],
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "required": true,
          "schema": { "type": "integer" },
          "description": "ID of the notification to delete"
        }
      ],
      "responses": {
        "200": {
          "description": "Notification deleted successfully",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string", "example": "Notification deleted successfully" }
                }
              }
            }
          }
        },
        "404": {
          "description": "Notification not found"
        },
        "500": {
          "description": "Internal server error"
        }
      }
    }
  },
  "/api/notification/{id}/mark-as-read": {
    "put": {
      "summary": "Mark a notification as read",
      "description": "Updates the status of a notification to 'read'",
      "tags": ["Notifications"],
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "required": true,
          "schema": { "type": "integer" },
          "description": "ID of the notification"
        }
      ],
      "responses": {
        "200": {
          "description": "Notification marked as read",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string", "example": "Notification marked as read" },
                  "notification": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer", "example": 1 },
                      "status": { "type": "string", "example": "read" }
                    }
                  }
                }
              }
            }
          }
        },
        "404": {
          "description": "Notification not found"
        },
        "500": {
          "description": "Internal server error"
        }
      }
    }
  },
  "/api/notification/system-downtime": {
    "post": {
      "summary": "Send system downtime notification",
      "description": "Sends a system downtime notification to all providers",
      "tags": ["Notifications"],
      "requestBody": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Scheduled maintenance at 2 AM" }
              }
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "System downtime notifications sent",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string", "example": "System downtime notifications sent successfully to all providers" },
                  "notifications": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer", "example": 1 },
                        "provider_id": { "type": "integer", "example": 101 },
                        "message": { "type": "string", "example": "Scheduled maintenance at 2 AM" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "404": {
          "description": "No providers found"
        },
        "500": {
          "description": "Internal server error"
        }
      }
    }
  }
    },
    "components": {
...
      "securitySchemes": {}
    }
  }